defobject: the most useful tool from my batteries suite.
Lots of times you just want to have a CLOS class, but not want to worry
about the exact finangling of the accessors. You just want a POD, doggone it.

Okay, well, maybe just us refugees from other languages. But still. Someone
wanted that idea, because there was a DEFSTRUCT. Too bad it's not a regular
CLOS object.

DEFOBJECT provides approximately the same interface as DEFSTRUCT, but
emits a CLOS object instead. You can inherit from it, specialize on it,
etc.  It's worked very well for me for quite a few months now, and I think
it will work well for the future.